# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Sukoon event management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  upcoming: 'upcoming',
  past: 'past',
  cancelled: 'cancelled',
} as const;

export interface Event {
  id: number;
  title: string;
  editionNumber: number;
  date: string;
  city: string;
  /** @nullable */
  venue: string | null;
  capacity: number;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  offerText?: string | null;
  status: EventStatus;
  /** @nullable */
  rsvpLink?: string | null;
  createdAt?: string;
}

export interface EventStats {
  eventId: number;
  capacity: number;
  confirmed: number;
  pending: number;
  waitlisted: number;
  declined: number;
  arrived: number;
}

export interface EventUpdateInput {
  price?: number;
  originalPrice?: number;
  offerText?: string;
}

export type AttendanceRequestStatus = typeof AttendanceRequestStatus[keyof typeof AttendanceRequestStatus];


export const AttendanceRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  declined: 'declined',
  waitlisted: 'waitlisted',
} as const;

export interface AttendanceRequest {
  id: number;
  eventId: number;
  name: string;
  phone: string;
  email: string;
  /** @nullable */
  socialHandle?: string | null;
  /** @nullable */
  heardAbout?: string | null;
  /** @nullable */
  mutualConnection?: string | null;
  /** @nullable */
  whyAttend?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  attendancePossibility?: string | null;
  status: AttendanceRequestStatus;
  /** @nullable */
  ticketCode?: string | null;
  checkedIn?: boolean;
  createdAt: string;
}

export interface AttendanceRequestInput {
  eventId: number;
  /** @minLength 2 */
  name: string;
  /** @minLength 10 */
  phone: string;
  email: string;
  socialHandle?: string;
  heardAbout?: string;
  mutualConnection?: string;
  whyAttend?: string;
  department?: string;
  attendancePossibility?: string;
}

export type RequestStatusUpdateStatus = typeof RequestStatusUpdateStatus[keyof typeof RequestStatusUpdateStatus];


export const RequestStatusUpdateStatus = {
  approved: 'approved',
  declined: 'declined',
  waitlisted: 'waitlisted',
} as const;

export interface RequestStatusUpdate {
  status: RequestStatusUpdateStatus;
}

export type GuestStatus = typeof GuestStatus[keyof typeof GuestStatus];


export const GuestStatus = {
  confirmed: 'confirmed',
  waitlisted: 'waitlisted',
} as const;

export interface Guest {
  id: number;
  requestId: number;
  eventId: number;
  name: string;
  phone: string;
  email: string;
  ticketCode: string;
  checkedIn: boolean;
  /** @nullable */
  checkedInAt?: string | null;
  status: GuestStatus;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminSession {
  username: string;
  name: string;
}

export interface UserSignUpInput {
  /** @minLength 2 */
  name: string;
  department: string;
  /** @minLength 10 */
  phone: string;
  email: string;
  /** @minLength 4 */
  password: string;
}

export interface UserLoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  department: string;
  phone: string;
  email: string;
}

export type UserPassStatus = typeof UserPassStatus[keyof typeof UserPassStatus];


export const UserPassStatus = {
  pending: 'pending',
  approved: 'approved',
  declined: 'declined',
  waitlisted: 'waitlisted',
} as const;

export interface UserPass {
  id: number;
  eventId: number;
  eventTitle: string;
  eventDate: string;
  /** @nullable */
  eventVenue: string | null;
  status: UserPassStatus;
  /** @nullable */
  ticketCode?: string | null;
  name: string;
  /** @nullable */
  department: string | null;
  /** @nullable */
  attendancePossibility?: string | null;
  checkedIn?: boolean;
  createdAt: string;
}

export type ListRequestsParams = {
eventId?: number;
status?: ListRequestsStatus;
};

export type ListRequestsStatus = typeof ListRequestsStatus[keyof typeof ListRequestsStatus];


export const ListRequestsStatus = {
  pending: 'pending',
  approved: 'approved',
  declined: 'declined',
  waitlisted: 'waitlisted',
} as const;

export type ListGuestsParams = {
eventId?: number;
};