# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sukoon event management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List all attendance requests for an event (admin only)
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, declined, waitlisted]
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRequest"
        "401":
          description: Unauthorized
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit an attendance request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceRequestInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRequest"
        "400":
          description: Validation error
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a single attendance request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRequest"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /requests/{id}/status:
    patch:
      operationId: updateRequestStatus
      tags: [requests]
      summary: Approve, decline, or waitlist a request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestStatusUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRequest"
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized
...
  /admin/events/{id}:
    patch:
      operationId: updateEvent
      tags: [admin]
      summary: Update event pricing and offers (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdateInput"
      responses:
        "200":
          description: Updated successfully
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all registered user accounts (admin only)
      responses:
        "200":
          description: List of user accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /admin/users/{id}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a registered user account (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account deleted successfully
        "401":
          description: Unauthorized
        "404":
          description: User not found
components:
  schemas:
...
    EventUpdateInput:
      type: object
      properties:
        price:
          type: number
        originalPrice:
          type: number
        offerText:
          type: string
    AttendanceRequest:
      type: object
      required: [id, eventId, name, phone, email, status, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        socialHandle:
          type: ["string", "null"]
        heardAbout:
          type: ["string", "null"]
        mutualConnection:
          type: ["string", "null"]
        whyAttend:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        attendancePossibility:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, declined, waitlisted]
        ticketCode:
          type: ["string", "null"]
        checkedIn:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AttendanceRequestInput:
      type: object
      required: [eventId, name, phone, email]
      properties:
        eventId:
          type: integer
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 10
        email:
          type: string
        socialHandle:
          type: string
        heardAbout:
          type: string
        mutualConnection:
          type: string
        whyAttend:
          type: string
        department:
          type: string
        attendancePossibility:
          type: string
    RequestStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, declined, waitlisted]
...
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [username, name]
      properties:
        username:
          type: string
        name:
          type: string
...
    UserProfile:
      type: object
      required: [id, name, department, phone, email]
      properties:
        id:
          type: integer
        name:
          type: string
        department:
          type: string
        phone:
          type: string
        email:
          type: string
...