# lib/api-zod/src/generated/types/attendanceRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Sukoon event management API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceRequestStatus } from './attendanceRequestStatus';

export interface AttendanceRequest {
  id: number;
  eventId: number;
  name: string;
  phone: string;
  email: string;
  /** @nullable */
  socialHandle?: string | null;
  /** @nullable */
  heardAbout?: string | null;
  /** @nullable */
  mutualConnection?: string | null;
  /** @nullable */
  whyAttend?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  attendancePossibility?: string | null;
  status: AttendanceRequestStatus;
  /** @nullable */
  ticketCode?: string | null;
  checkedIn?: boolean;
  createdAt: Date;
}