# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Sukoon event management API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  name: string;
  department: string;
  phone: string;
  email: string;
}